requirements.txt
alembic==1.9.4
aniso8601==9.0.1
boto3==1.28.40
botocore==1.31.40
CacheControl==0.12.11
cachetools==5.3.0
certifi==2022.12.7
celery==5.3.1
charset-normalizer==3.1.0
click==8.1.3
firebase-admin==6.1.0
Flask==2.3.2
Flask-Cors==3.0.10
Flask-HTTPAuth==4.7.0
Flask-JWT-Extended==4.4.4
flask-restx==1.1.0
Flask-SQLAlchemy==3.0.2
google-api-core==2.11.0
google-api-python-client==2.86.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-cloud-core==2.3.2
google-cloud-firestore==2.11.1
google-cloud-storage==2.9.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.59.0
greenlet==2.0.1
grpcio==1.54.2
grpcio-status==1.48.2
httplib2==0.22.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.17.3
Mako==1.2.4
MarkupSafe==2.1.1
msgpack==1.0.5
proto-plus==1.22.2
protobuf==3.20.3
psycopg2~=2.9.9
pyasn1==0.5.0
pyasn1-modules==0.3.0
pydantic==1.10.5
pydantic[email]
PyJWT==2.6.0
pyotp==2.8.0
pyparsing==3.0.9
python-crontab==3.0.0
requests==2.30.0
rsa==4.9
six==1.16.0
SQLAlchemy==1.4.44
sqlalchemy-filters
uritemplate==4.1.1
urllib3==1.26.16
Werkzeug==2.3.4
redis
aiohttp
asyncio